fix scan_ext warning

+++ src/sensor/scan_ext.h
@@ -25,6 +25,6 @@
 
 #include "interface.h"
 
-int sensor_scan_ext(const sensor_ext_ssi_t *ext_ssi, uint8_t *ext_dev_addr, uint8_t *ext_dev_reg, int dev_addr_count, const uint8_t dev_addr[], const uint8_t dev_reg[], const uint8_t dev_id[], const int dev_ids[]);
+int sensor_scan_ext(const sensor_ext_ssi_t *ext_ssi, uint16_t *ext_dev_addr, uint8_t *ext_dev_reg, int dev_addr_count, const uint8_t dev_addr[], const uint8_t dev_reg[], const uint8_t dev_id[], const int dev_ids[]);
 
 #endif
